+++ server/models/student.go
@@ -9,7 +9,7 @@ import (
 type Student struct {
 	User `bun:",inherit"`
 
-	Score    int
+	Score    int8
 	Accepted bool
 
 	EventsRegistered []Event  `bun:"m2m:student_to_events,join:Student=Event"`
@@ -21,10 +21,10 @@ func (Student) IsUser() {}
 type StudentToEvent struct {
 	StudentMail string   `bun:",pk"`
 	Student     *Student `bun:"rel:belongs-to,join:student_mail=mail"`
-	EventID     int      `bun:",pk"`
+	EventID     int32    `bun:",pk"`
 	Event       *Event   `bun:"rel:belongs-to,join:event_id=id"`
 	RoomNumber  string   `bun:",pk"`
-	BuildingID  int      `bun:",pk"`
+	BuildingID  int32    `bun:",pk"`
 	Room        *Room    `bun:"rel:belongs-to,join:room_number=number,join:building_id=building_id"`
 }
